(Simple) iterative deepening depth-first search
A simple and modular implementation of IDDFS as an FPC. The basic
idea is as follows: the certificate's only responsibility is to
guide proof search following the IDDFS discipline adapted to the
level of granularity natively supported by the kernels of the FPC
framework: that of bipoles. The present implementation makes no
assumptions on computational bounds whatsoever and proceeds by
bootstrapping the process with an initial certificate constructor,
which is responsible for immediately generating the properly
iterative certificate and inductively generating the natural search
bounds associated with it. In this fashion, each unit enables one
instance of the release rule along a certain branch, proceeding
until all branches are closed or the kernel backtracks to the
bootstrapper, which then increments the bound and restarts the
process.

Clearly, this requires an atomic judgment to be added to the logic
of the clerk or expert that initiates search. Currently, the type
'nonneg' as defined in 'atomic' is used, an a temporary predicate
for typing judgment added there (of course, there is the assumption
that other modules will not abuse the open world assumption, and
leave the definition unaltered, lest hilarity ensue). Ideally,
further modularity would see common types exported to their own
modules and reused accordingly, at the cost of more interesting
module composition.

Described here is the "unbounded" variant of IDDFS. If a maximum
depth is desired, the process is simpler: it suffices to provide
one single certificate type with this bound and to let all clerks
and experts pass the certificate along, unchanged, with the sole
exception of the release experts, responsible for decrementing the
bound (again, with some minor bookkeeping to increment the
effective bound until it hits the maximum allowed on the part of
the bootstrappers). This approach has not been followed here.

An implementation alternative is to implement this transformation
using unmarshalling code.

The usefulness of this FPC is closely related to pairing. By
itself, it cannot constrain nondeterministic behavior enough to
guide proof search (specifically, unfoldings are completely
unchecked). However, it offers a very simple control mechanism
which can used to complement the semantics of other FPCs, as
illustrated by its combination with the simple atomic outlines,
composed in 'atomic-iddfs'. A Bedwyr translation with its
corresponding test harness is given to illustrate the approach. It
should be noted that atomic outlines, when paired in this way, can
dispense with any and all bounds on the number of bipoles to
explore without fear: the IDDFS half controls this dimension of
progress effectively.

The current version implements bootstrapping only for the universal
clerk. For many (most?) common theorems, this will suffice, but in
the general case a more flexible treatment must be finished. If one
assumes checking to commence with an asynchronous phase, only
clerks would be affected.

A final point of interest is that the addition of inductive typing
judgments complicates translation in the following two senses:

 - There appear dependencies between Bedwyr definitions, which need
   to be ordered in a suitable way (forward declarations are not
   supported).

 - Furthermore, the order of the clauses within a single inductive
   definitions will be significant and must be correctly
   established in lambda Prolog and respected by the translation.

The current translation from lambda Prolog to Bedwyr via Abella
does not ensure that these constraints are fulfilled.