Remove vestigial external indexes
An internal singleton index type is created for dummy purposes,
removing the last external dependency in the IDDFS module. For
simplicity, all instances of internal indexes are replaced with
this dummy, even though a partial substitution was feasible as
illustrated by previous versions. Critically, and due to
unification constraints given by L_\lambda in Bedwyr, at least one
"side" of each write-read operation on indexes needs to be defined
for checking to succeed, even if there is no information involved
in the process.