lProlog proof of concept for clerks and experts
The main branching experts are defined, and some precisions and
possible simplifications noted along with the .mod file, where most
predicates are commented out until further notice. The simplest and
most insufficient simulation of randomness is mocked in a way that
hopefully will be portable to Bedwyr without major changes to
explore proof search in the kernel.