Initial Bedwyr translation of quick
Adapting the Abella-based translator to parse generic lProlog is
more trouble than it's worth at this point (among other things, it
cannot cover predefined types such as lists, or general code blocks
including but not limited to I/O and mapping between families of
predefined predicates).

An initial translation and tests are created to treat the
generation of simple inductive types, for now simple integers. The
resulting code is greatly simplified in its more subtle
(projected) pseudorandom behavior, but it serves to illustrate
Bedwyr's new interactive capabilities. Everything seems to work
well, with an interesting effect. To simulate random sampling a
"toin coss" is requested interactively and the branch to take
decided based on its binary value. As it stands, a 0 value works as
expected, whereas a 1 value must be entered twice to succeed.
Otherwise, everything seems to work as expected. It would be
interesting to have certificates contain the generated (polymorphic)
term.

The equality initial rule is added both to lProlog and Bedwyr.