Kernel calls orExpert exactly once
Previously, the kernel called orExpert once for each branch
unifying with the corresponding choice. If the expert is functional
there is no issue except perhaps one of efficiency. When side
effects are added to the mix, be it randomness, user interaction,
etc., this may lead to loss of those effects. The solution is to
call the expert exactly once and unify after the choice, if any,
along with any side effects leading to the choice, is made.