Add structure for tests on involutivity of reverse
A typing judgment on a specific type of list, namely lists of
natural numbers, is defined along with a fixed point designed to
accommodate a QuickCheck-like testing structure. The appropriate,
two-stage certificate structure is currently missing, but will
consist of a generation phase followed by a checking phase, to be
implemented next.