Simply two-stage quick certificates
The QuickCheck-like process implemented by quick certificates is
refined to take place in two stages:

 1. A first round of generation of random instances of inductive
    types.

 2. A second round of computation based on those generated values,
    in an attempt to falsify the property.

To do this, a set of partial extensions are made to certificate
constructors and the clerks and experts processing those. (The
encoding is assuredly suboptimal, but is readable and flexible
enough to try new things.) Essentially, the initial certificate
would be a pair of certificates corresponding to those two phases,
proceeding as the original definitions for the generation phase
(namely, focus on the right, introduce eigenvariables and proceed
until the first branching point). The first branching point is
expected to be a conjunction separating both phases and splitting
the certificate in its constituents. The first branch is assumed
to be that of generation proper, taking place as in previous
versions. The second branch is that of computation; for this a
very simple certificate constructor allowing unlimited unfolding
on the right, and some final equality processing on release, is
presented. These definitions are certainly not exhaustive, but
should cover common cases (and be extensible as needed).

It would be nice to have a modular definition of these new parts,
e.g. in lProlog, and combine phases in a more flexible way,
although there are limitations imposed by the new complexities of
the generation phase and how their differential and imperfect
handling by both lProlog and Bedwyr.