Driver for random number generation in Bedwyr
A simple script to automate the generation of random choices
interfaces directly with Bedwyr's I/O, as encoded in the quick FPC.
It makes use of bash's rudimentary capabilities to prototype a
means of overcoming the interfacing limitations of the model
checking system.

Obviously, the hack is sensitive to changes in formatting and
unexpected input, as well as limited resolution in pseudorandom
number generation and the robustness thereof.

The weights in quick certificates finally become relevant, and must
be furthermore output by the corresponding expert in order for the
user (or the script, as the case may be) be capable of taking these
numbers into account when making their choice.