Example: insertion on ordered lists
A somewhat more interesting use case of QuickCheck showing how the
invariant of preservation of ordering cannot be falsified under a
correct implementation, but it can under a buggy one. Ideally, one
will want to retrieve those instances that falsify the properties
of interest.

To account for this more interesting (and realistic) usage, the FPC
needs to be extended to perform slightly more extensive search,
including some asynchronous unfolding on the left, with the
corresponding risk of nontermination, which can be controlled
externally by pairing with a limiter certificate.