abella: syntax for sim-automata-harness-simple
Standard list of changes.

Positive assertions work without issue.

Negative assertions would require a modified version of the search
tactic, which in the event of failure informs of whether all paths
finished with failure, or some were suspended for lack of depth. In
the former case, we can safely conclude that the property is
absolutely not provable.