abella: temporary test harness
Changes in indexing made to accommodate the lack of I/O facilities
in Abella interfere with lemma names as strings. For this reason,
and until the admin FPC has been adapted to the new conventions, the
test harness only invokes those examples that work on the simple FPC
(which has been fully adapted by now).