abella: Bedwyr-to-Abella preprocessor
This simple preprocessor script translates compatible Bedwyr files
into their Abella equivalents, by virtue of preprocessing #include
directives recursively, and keeping track of files that have
already been included, so as not to include them more than once. It
makes no effort to compensate for Abella's limitations, so that
Bedwyr files must themselves be compatible; the preprocessing only
circumvents the problem of modularity and imports by composing the
full module in much the same way Bedwyr does.

In principle, one will call this preprocessor on a harness file,
which orchestrates a complete instance of the FPCcheck kernel, with
FPC, definitions and theorems. (Currently, this would be limited to
examples/*-harness-simple.thm files). This file, passed as a
parameter, will be preprocessed and incorporated into the output,
which can be directly fed to Abella.

Directives other than #include are not processed and will cause
errors in Abella (assuming the files are valid Bedwyr). In
particular, #assert and #assert_not directives need to be
translated into Theorem statements, with the following adjustments:

 - If there are any anonymous variables, replace them with standard
   logic variables, with names compatible with both Bedwyr and
   Abella.

 - Any logical variables must be included in an existential
   preamble in the theorem. For example:

     #assert some_property _ _.

   becomes

     Theorem t : exists X Y, some_property X Y.

 - For each theorem, a proof script should be given. If provable,
   a simple search command with an adequate depth bound should
   suffice.

For all this to work, a version of Abella recent enough to
incorporate the fix for issue #56 must be used.