quickcheck: exhaustive generation based on size
A second type of generator is based on the size of the derivation tree
for the instance, i.e., the number of constructors in the generated
term. A possible encoding of this strategy uses difference lists,
written with the first component given as the exact desired number of
constructores, and the second initialized to zero (this is more detail
than is strictly necessary, and eventually could and should be hidden
from the user).

In this scheme, unfolding works as expected, decrementing the bound in
the first component (and propagating the second). Equality as the end of
a branch in the generation assigns the unused bound to the second
component. Conjunction splits the bound in two in the following way: the
"incoming bound" becomes the incoming of the left subformula, and the
"outgoing bound" is assigned to a new, intermediate variable, which
becomes the incoming (remaining) bound for the right subformula, whose
outgoing bound must in turn be the original outgoing bound, i.e., in
this case, zero, corresponding to instances making use of exactly the
number of constructors given in the first component.

Examples are shown with the minimum bound that exhibits the properties
under consideration, and for the negative case, a somewhat larger bound.
Note that in the cases given here a larger bound, even used
exhaustively, guarantees the existence of counterexamples for all values
above the minimum, although this can be subtle and an exhaustive
exploration comprising all bounds up to and including the given number
is, generally, more useful and conceptually easier to understand,
although it can be substantially costlier to apply.